refactor: rename github-issue-creator.lib.mjs to github-error-reporter.lib.mjs
Address PR feedback: the file name "github-issue-creator" was misleading since
the module handles error reporting via GitHub (issues and comments), not just
issue creation. Renamed to "github-error-reporter" for clarity.

Updated all imports and references in:
- src/solve.error-handlers.lib.mjs
- experiments/test-issue-306-auto-issue-creator.mjs
- docs/case-studies/issue-1462/README.md
- .changeset/fix-pr-creation-error-messaging.md
