+++ .changeset/fix-pr-creation-error-messaging.md
@@ -8,3 +8,4 @@ fix: improve PR creation failure error messaging and log upload fallback (Issue 
 - Upload failure logs to the issue as fallback when PR is not available (--attach-logs)
 - Capture and log `gh pr create` stdout/stderr in verbose mode for root cause diagnosis
 - Add fallback GitHub user detection via `gh auth status` when `gh api user` fails
+- Rename `github-issue-creator.lib.mjs` to `github-error-reporter.lib.mjs` for clarity
